updates to canonical mapper
