+++ presidio_evaluator/entity_mapping/__init__.py
@@ -1,28 +1,26 @@
-"""Entity mapping: resolve raw entity labels to canonical EntityHierarchy entities."""
+"""Entity mapping: resolve raw entity labels to canonical entities."""
 
 from presidio_evaluator.entity_mapping.mapper import (
-    CanonicalMapper,
-    EntityMapper,
-    IncompleteMapping,
-)
-
-from presidio_evaluator.entity_mapping.hierarchy import (
     ALL_CANONICAL_ENTITIES,
     CANONICAL_TO_BRANCH,
-    COUNTRIES,
-    HIERARCHY,
     RAW_TO_CANONICAL,
+    CanonicalMapper,
     EntityHierarchy,
-    EntityNotMappedError,
+    IncompleteMapping,
     canonicalize,
     get_branch,
     print_hierarchy,
 )
 
+from presidio_evaluator.entity_mapping.definitions import (
+    COUNTRIES,
+    HIERARCHY,
+    EntityNotMappedError,
+)
+
 __all__ = [
     # Mapper
     "CanonicalMapper",
-    "EntityMapper",
     "IncompleteMapping",
     # Entity hierarchy
     "ALL_CANONICAL_ENTITIES",
